Allow for custom public exponent value in keygen
This commit adds a function to `rsa::algorithms` called
`generate_multi_prime_key_with_exp` which allows the caller
to specify a custom value for the public key exponent.

This commit also adds a convenience routine to `rsa::RSAPrivateKey`
called `new_with_exp` which allows the caller to specify the
custom value for the public key exponent as part of `rsa::RSAPrivateKey`
constructor.

Exposing the public key exponent matches an OpenSSL call
`openssl::rsa::generate_with_e` which is useful in certain
settings such when generating the signing keys for SGX enclaves.