Add `RsaPrivateKey::crt_coefficient` (#158)
This method is useful for key serialization formats and is internally
useful within the context of the PKCS#1/PKCS#8 implementations.

Externally I'm working on adding support to the `ssh-key` crate. I could
implement this same thing by pulling in `num_bigint_dig::ModInverse`,
it'd be a lot easier if this purpose-dedicated method were available.